rename ciphers.loader -> ciphers.cipher
