+++ lib/ansible/parsing/vault/ciphers/cipher.py
